move arc_2d to it's own file
(in another MR i'll write a 3d arc function)
